Add files via upload
this code should be used for the base of a bigger project, or to show how to add a menu bar.  The Microsoft instructions are ambiguous, actually to be blunt they are terrible.   AND anyone who says otherwise is a seasoned programmer defending MicroSoft, condescending,  and is lying to you.